+++ Src/FluentAssertions/Specialized/AsyncFunctionAssertions.cs
@@ -11,21 +11,17 @@ namespace FluentAssertions.Specialized
     /// Contains a number of methods to assert that an asynchronous method yields the expected result.
     /// </summary>
     [DebuggerNonUserCode]
-    public class AsyncFunctionAssertions : DelegateAssertions<Func<Task>>
+    public class AsyncFunctionAssertions<TTask, TAssertions> : DelegateAssertions<Func<TTask>, TAssertions>
+        where TTask : Task
+        where TAssertions : AsyncFunctionAssertions<TTask, TAssertions>
     {
-        public AsyncFunctionAssertions(Func<Task> subject, IExtractExceptions extractor) : this(subject, extractor, new Clock())
+        public AsyncFunctionAssertions(Func<TTask> subject, IExtractExceptions extractor) : this(subject, extractor, new Clock())
         {
         }
 
-        public AsyncFunctionAssertions(Func<Task> subject, IExtractExceptions extractor, IClock clock) : base(subject, extractor, clock)
+        public AsyncFunctionAssertions(Func<TTask> subject, IExtractExceptions extractor, IClock clock) : base(subject, extractor, clock)
         {
-            Subject = subject;
         }
-
-        /// <summary>
-        /// Gets the <see cref="Func{Task}"/> that is being asserted.
-        /// </summary>
-        public new Func<Task> Subject { get; }
 
         protected override string Identifier => "async function";
 
@@ -34,6 +30,77 @@ namespace FluentAssertions.Specialized
         protected override void InvokeSubject()
         {
             Subject.ExecuteInDefaultSynchronizationContext().GetAwaiter().GetResult();
+        }
+
+        /// <summary>
+        /// Asserts that the current <typeparamref name="TTask"/> will complete within specified time.
+        /// </summary>
+        /// <param name="timeSpan">The allowed time span for the operation.</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public AndConstraint<TAssertions> CompleteWithin(
+            TimeSpan timeSpan, string because = "", params object[] becauseArgs)
+        {
+            Execute.Assertion
+                .ForCondition(Subject is object)
+                .BecauseOf(because, becauseArgs)
+                .FailWith("Expected {context:task} to complete within {0}{reason}, but found <null>.", timeSpan);
+
+            TTask task = Subject.ExecuteInDefaultSynchronizationContext();
+            bool completed = Clock.Wait(task, timeSpan);
+
+            Execute.Assertion
+                .ForCondition(completed)
+                .BecauseOf(because, becauseArgs)
+                .FailWith("Expected {context:task} to complete within {0}{reason}.", timeSpan);
+
+            return new AndConstraint<TAssertions>((TAssertions)this);
+        }
+
+        /// <summary>
+        /// Asserts that the current <typeparamref name="TTask"/> will complete within the specified time.
+        /// </summary>
+        /// <param name="timeSpan">The allowed time span for the operation.</param>
+        /// <param name="because">
+        /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
+        /// is needed. If the phrase does not start with the word <i>because</i>, it is prepended automatically.
+        /// </param>
+        /// <param name="becauseArgs">
+        /// Zero or more objects to format using the placeholders in <paramref name="because" />.
+        /// </param>
+        public async Task<AndConstraint<TAssertions>> CompleteWithinAsync(
+            TimeSpan timeSpan, string because = "", params object[] becauseArgs)
+        {
+            Execute.Assertion
+                .ForCondition(Subject is object)
+                .BecauseOf(because, becauseArgs)
+                .FailWith("Expected {context:task} to complete within {0}{reason}, but found <null>.", timeSpan);
+
+            using (var timeoutCancellationTokenSource = new CancellationTokenSource())
+            {
+                TTask task = Subject.ExecuteInDefaultSynchronizationContext();
+
+                Task completedTask =
+                    await Task.WhenAny(task, Clock.DelayAsync(timeSpan, timeoutCancellationTokenSource.Token));
+
+                if (completedTask == task)
+                {
+                    timeoutCancellationTokenSource.Cancel();
+                    await completedTask;
+                }
+
+                Execute.Assertion
+                    .ForCondition(completedTask == task)
+                    .BecauseOf(because, becauseArgs)
+                    .FailWith("Expected {context:task} to complete within {0}{reason}.", timeSpan);
+
+                return new AndConstraint<TAssertions>((TAssertions)this);
+            }
         }
 
         /// <summary>
@@ -108,7 +175,7 @@ namespace FluentAssertions.Specialized
         /// <param name="becauseArgs">
         /// Zero or more objects to format using the placeholders in <paramref name="because" />.
         /// </param>
-        public async Task NotThrowAsync(string because = "", params object[] becauseArgs)
+        public async Task<AndConstraint<TAssertions>> NotThrowAsync(string because = "", params object[] becauseArgs)
         {
             Execute.Assertion
                 .ForCondition(Subject is object)
@@ -123,6 +190,8 @@ namespace FluentAssertions.Specialized
             {
                 NotThrow(exception, because, becauseArgs);
             }
+
+            return new AndConstraint<TAssertions>((TAssertions)this);
         }
 
         /// <summary>
@@ -135,7 +204,7 @@ namespace FluentAssertions.Specialized
         /// <param name="becauseArgs">
         /// Zero or more objects to format using the placeholders in <paramref name="because" />.
         /// </param>
-        public async Task NotThrowAsync<TException>(string because = "", params object[] becauseArgs)
+        public async Task<AndConstraint<TAssertions>> NotThrowAsync<TException>(string because = "", params object[] becauseArgs)
             where TException : Exception
         {
             Execute.Assertion
@@ -151,6 +220,8 @@ namespace FluentAssertions.Specialized
             {
                 NotThrow<TException>(exception, because, becauseArgs);
             }
+
+            return new AndConstraint<TAssertions>((TAssertions)this);
         }
 
         /// <summary>
@@ -176,7 +247,7 @@ namespace FluentAssertions.Specialized
         /// Zero or more objects to format using the placeholders in <paramref name="because" />.
         /// </param>
         /// <exception cref="ArgumentOutOfRangeException">Throws if waitTime or pollInterval are negative.</exception>
-        public Task NotThrowAfterAsync(TimeSpan waitTime, TimeSpan pollInterval, string because = "", params object[] becauseArgs)
+        public Task<AndConstraint<TAssertions>> NotThrowAfterAsync(TimeSpan waitTime, TimeSpan pollInterval, string because = "", params object[] becauseArgs)
         {
             if (waitTime < TimeSpan.Zero)
             {
@@ -198,7 +269,7 @@ namespace FluentAssertions.Specialized
 
             return assertionTask();
 
-            async Task assertionTask()
+            async Task<AndConstraint<TAssertions>> assertionTask()
             {
                 TimeSpan? invocationEndTime = null;
                 Exception exception = null;
@@ -209,7 +280,7 @@ namespace FluentAssertions.Specialized
                     exception = await InvokeWithInterceptionAsync(wrappedSubject);
                     if (exception is null)
                     {
-                        return;
+                        return new AndConstraint<TAssertions>((TAssertions)this);
                     }
 
                     await Clock.DelayAsync(pollInterval, CancellationToken.None);
@@ -219,6 +290,8 @@ namespace FluentAssertions.Specialized
                 Execute.Assertion
                     .BecauseOf(because, becauseArgs)
                     .FailWith("Did not expect any exceptions after {0}{reason}, but found {1}.", waitTime, exception);
+
+                return new AndConstraint<TAssertions>((TAssertions)this);
             }
         }